+++ src/base/ApplicationECSService.ts
@@ -66,6 +66,7 @@ export class ApplicationECSService extends Resource {
   public readonly codeDeployApp?: ApplicationECSAlbCodeDeploy;
   public readonly ecrRepos: ecr.EcrRepository[];
   public readonly taskDefinition: ecs.EcsTaskDefinition;
+  public ecsIam: ApplicationECSIAM;
   private readonly config: ApplicationECSServiceProps;
 
   constructor(
@@ -311,8 +312,6 @@ export class ApplicationECSService extends Resource {
           cidrBlocks: [],
           ipv6CidrBlocks: [],
           prefixListIds: [],
-          // @ts-ignore: https://github.com/hashicorp/terraform-cdk/issues/282
-          self: false,
         },
       ];
     }
@@ -327,8 +326,6 @@ export class ApplicationECSService extends Resource {
         ipv6CidrBlocks: [],
         prefixListIds: [],
         securityGroups: [],
-        // @ts-ignore: https://github.com/hashicorp/terraform-cdk/issues/282
-        self: false,
       },
     ];
 
@@ -398,7 +395,7 @@ export class ApplicationECSService extends Resource {
       containerDefs.push(buildDefinitionJSON(def));
     });
 
-    const ecsIam = new ApplicationECSIAM(this, 'ecs-iam', {
+    this.ecsIam = new ApplicationECSIAM(this, 'ecs-iam', {
       prefix: this.config.prefix,
       tags: this.config.tags,
       taskExecutionDefaultAttachmentArn:
@@ -414,8 +411,8 @@ export class ApplicationECSService extends Resource {
       // why are container definitions just JSON? can we get a real construct? sheesh.
       containerDefinitions: `[${containerDefs}]`,
       family: `${this.config.prefix}`,
-      executionRoleArn: ecsIam.taskExecutionRoleArn,
-      taskRoleArn: ecsIam.taskRoleArn,
+      executionRoleArn: this.ecsIam.taskExecutionRoleArn,
+      taskRoleArn: this.ecsIam.taskRoleArn,
       cpu: this.config.cpu.toString(),
       memory: this.config.memory.toString(),
       requiresCompatibilities: ['FARGATE'],